# day1_todo_app/internal/infra/datastore/model/todos.gen.go
// Code generated by gorm.io/gen. DO NOT EDIT.
// Code generated by gorm.io/gen. DO NOT EDIT.
// Code generated by gorm.io/gen. DO NOT EDIT.

package model

import (
	"time"
)

const TableNameTodo = "todos"

// Todo ToDo
type Todo struct {
	ID          int64      `gorm:"column:id;type:bigint;primaryKey;autoIncrement:true;comment:ToDo ID" json:"id"`                                                                       // ToDo ID
	UserID      int64      `gorm:"column:user_id;type:bigint;not null;index:idx_user_archived_sort,priority:1;index:idx_user_created,priority:1;comment:ãƒ¦ãƒ¼ã‚¶ãƒ¼ID" json:"user_id"` // ãƒ¦ãƒ¼ã‚¶ãƒ¼ID
	Title       string     `gorm:"column:title;type:varchar(255);not null;comment:ã‚¿ã‚¤ãƒˆãƒ«" json:"title"`                                                                           // ã‚¿ã‚¤ãƒˆãƒ«
	Description *string    `gorm:"column:description;type:text;comment:è©³ç´°" json:"description"`                                                                                      // è©³ç´°
	Status      string     `gorm:"column:status;type:enum('not started','in progress','done','pending','cancel');not null;default:not started;comment:çŠ¶æ…‹" json:"status"`            // çŠ¶æ…‹
	SortOrder   float64    `gorm:"column:sort_order;type:double;not null;index:idx_user_archived_sort,priority:3;comment:ã‚½ãƒ¼ãƒˆé †" json:"sort_order"`                               // ã‚½ãƒ¼ãƒˆé †
	CreatedAt   *time.Time `gorm:"column:created_at;type:timestamp;index:idx_user_created,priority:2;default:CURRENT_TIMESTAMP;comment:ä½œæˆæ—¥æ™‚" json:"created_at"`                 // ä½œæˆæ—¥æ™‚
	UpdatedAt   time.Time  `gorm:"column:updated_at;type:datetime;not null;default:CURRENT_TIMESTAMP;comment:æ›´æ–°æ—¥æ™‚" json:"updated_at"`                                           // æ›´æ–°æ—¥æ™‚
	ArchivedAt  *time.Time `gorm:"column:archived_at;type:timestamp;index:idx_user_archived_sort,priority:2;comment:ã‚¢ãƒ¼ã‚«ã‚¤ãƒ–æ—¥æ™‚" json:"archived_at"`                          // ã‚¢ãƒ¼ã‚«ã‚¤ãƒ–æ—¥æ™‚
}

// TableName Todo's table name
func (*Todo) TableName() string {
	return TableNameTodo
}